Added transparent material, Car rotates in place at endpoint, escape to exit mode
I tried for a couple hours to get the rotate to work in a fancy way, like parking, this however, was very unsuccessful, to the point I just implemented a turn on location to fix it.

There is now a transparent material that gets applied to the full car, so it looks like a ghost placement instead of just another car.

Escape is now a viable option to quit the placement mode.